requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
openai>=1.0.0
elevenlabs==0.2.26
spotipy==2.23.0
python-multipart==0.0.6
jinja2==3.1.2
pydantic==2.5.2
alembic==1.12.1
psycopg2-binary==2.9.9
